refactor(api): reorganize user creation feature with feature-based folder structure
This commit refactors the codebase organization for the user creation feature
by moving from a type-based structure (implementations/interfaces) to a
feature-based structure.

Changes:
- Removed old structure:
  - Deleted application/use-cases/user/implementations/create-user.ts
  - Deleted application/use-cases/user/implementations/index.ts
  - Deleted application/use-cases/user/interfaces/create-user.ts
  - Deleted application/use-cases/user/interfaces/index.ts

- Added new feature-based structure:
  - Created application/use-cases/user/create-user/create-user.ts
  - Created application/use-cases/user/create-user/create-user.interface.ts
  - Created application/use-cases/user/create-user/index.ts

- Updated imports in application/use-cases/user/index.ts to reflect the new organization

This change aims to improve code maintainability by grouping related files
based on their feature rather than their type, making it easier to navigate
and update related code.
