+++ apps/express-app/src/application/use-cases/user/index.ts
@@ -1,2 +1,1 @@
-export * from "./implementations";
-export * from "./interfaces";
+export * from "./create-user";
