--- apps/express-app/src/application/use-cases/user/interfaces/index.ts
@@ -1,1 +1,0 @@
-export * from "./create-user";
